Reorganize files and adjust CSS for mobile
