+++ src/Articles/Story/Story.css
@@ -20,7 +20,7 @@
   }
 
   .tile-image {
-   height: 25%;
-   max-width: 700px;
+   height: 30%;
+   width: 100%
   }
 }
